# rms-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RMS API Documentation",
        "description": "API du Revenu Minimum Solidaire (RMS) : auth, portefeuille, file FIFO, rotations, timeline.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/wallet/adhesion": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Payer l'adhésion (simulation) et entrer dans la file FIFO",
                "operationId": "2aacf27e3a065c31f79b32c5416248b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Adhésion enregistrée, entrée dans la file"
                    },
                    "400": {
                        "description": "Utilisateur non éligible ou déjà en file"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wallet/remboursement": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Payer une tranche de remboursement (simulation)",
                "operationId": "b8d31152e3ed3a573515f3cb0607d4a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Remboursement enregistré"
                    },
                    "400": {
                        "description": "Erreur métier"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wallet/repayment": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Remboursement RMS en cours",
                "operationId": "5ec6ddd7257ba29e9ca67dc8f6f306a7",
                "responses": {
                    "200": {
                        "description": "Infos remboursement (ou null)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "État du wallet + 10 dernières transactions",
                "operationId": "c46516e6ae316194282761be57b0a915",
                "responses": {
                    "200": {
                        "description": "Wallet info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "number",
                                            "example": 150000
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "XOF"
                                        },
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}